% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{size}

\item{p}{probability}
}
\value{
barplot of relative frequency of successes per trial
}
\description{
Binomial Simulation
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
