% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the sample}

\item{fun}{the function to use}

\item{alpha}{1-confidence interval}

\item{cx}{scaling size for labels}

\item{...}{optional parameters for the histogram}
}
\value{
a list of the confidence interval, function, sample, and statistic
}
\description{
Title
}
\examples{
myboot2(x=c(1,1,1,2,2,2,2,3,3,3,4,4))
}
