% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootpval.R
\name{mybootpval}
\alias{mybootpval}
\title{Title}
\usage{
mybootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{sample vector}

\item{conf.level}{the confidence interval}

\item{iter}{number of iterations}

\item{mu0}{the null hypothesis}

\item{test}{type of test}
}
\value{
list of the pvalue, tcalc, sample size, sample, test type, confidence interval
}
\description{
Title
}
