% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Title}
\usage{
mycltu(n = 2, iter = 1000, a = 0, b = 10)
}
\arguments{
\item{n}{# sample size}

\item{iter}{# number of iterations}

\item{a}{# lower bound}

\item{b}{# upper bound}
}
\value{
#a vector w of sample means of a random uniform distribution
}
\description{
Title
}
\examples{
mycltu(n=2,iter=1000,a=0,b=10)
}
