% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the function to evaluate}

\item{x}{the distribution}

\item{param}{distribution parameter ex. lambda}

\item{...}{extra specifications for the plot}
}
\value{
list of maximum likelihood location, param at location, y at location, and slope
}
\description{
Title
}
\examples{
mymaxlik(x=y,param=seq(0,1,length=1000),lfun=logbin,main="Binomial,n=20",cex.main=1.5)
}
