% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard dev}

\item{a}{upper bound for probability}
}
\value{
figure with graph of function, shaded area, and calculated probability
}
\description{
Title
}
\examples{
myncurve(10,2,7)
}
